+++ core/cli/main.go
@@ -15,7 +15,9 @@ import (
 	"github.com/kurtosis-tech/kurtosis/commons/docker_manager"
 	"github.com/kurtosis-tech/kurtosis/commons/enclave_manager"
 	"github.com/kurtosis-tech/kurtosis/commons/enclave_manager/enclave_context"
+	"github.com/kurtosis-tech/kurtosis/commons/logrus_log_levels"
 	"github.com/kurtosis-tech/kurtosis/initializer/api_container_launcher"
+	"github.com/kurtosis-tech/kurtosis/initializer/docker_flag_parser"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
 	"golang.org/x/crypto/ssh/terminal"
@@ -23,6 +25,7 @@ import (
 	"io"
 	"math/rand"
 	"os"
+	"strings"
 	"time"
 )
 
@@ -31,19 +34,14 @@ const (
 	errorExitCode = 1
 
 	enclaveDataVolMountpointOnReplContainer = "/kurtosis-enclave-data"
-)
 
-const (
-	// TODO Read this from either:
-	//  1) the Kurt Core version if we're inside a testsuite repo
-	//  2) a global Kurtosis config if not
-	apiContainerImage = "kurtosistech/kurtosis-core_api:mieubrisse_enclave-creation-cli"
-
-	// TODO make this configurable somehow
-	kurtosisLogLevel = logrus.DebugLevel
-
-	// TODO make configurable
-	javascriptReplImage = "test-repl-image"
+	// TODO These defaults aren't great - it will just start a Kurtosis interactive with the latest
+	//  of both images, which may or may not be compatible - what we really need is a system that
+	//  detects what version of the API container/REPL to start based off the Kurt Core API version
+	// TODO It's also not great that these are hardcoded - they should be hooked into the build system,
+	//  to guarantee that they're compatible with each other
+	defaultApiContainerImage = "kurtosistech/kurtosis-core_api"
+	defaultJavascriptReplImage = "kurtosistech/javascript-interactive-repl"
 
 	shouldPublishPorts = true
 
@@ -53,7 +51,35 @@ const (
 	enclaveIdTimestampFormat = "2006-01-02T15.04.05"
 
 	isPartitioningEnabled = true
+
+	apiContainerImageArg = "kurtosis-api-image"
+	javascriptReplImageArg = "javascript-repl-image"
+	kurtosisLogLevelArg = "kurtosis-log-level"
 )
+var defaultKurtosisLogLevel = logrus.InfoLevel.String()
+var flagConfigs = map[string]docker_flag_parser.FlagConfig{
+	apiContainerImageArg: {
+		Required: false,
+		HelpText: "The image of the Kurtosis API container to use inside the enclave",
+		Default:  defaultApiContainerImage,
+		Type:     docker_flag_parser.StringFlagType,
+	},
+	javascriptReplImageArg: {
+		Required: false,
+		HelpText: "The image of the Javascript REPL to connect to the enclave with",
+		Default: defaultJavascriptReplImage,
+		Type: docker_flag_parser.StringFlagType,
+	},
+	kurtosisLogLevelArg: {
+		Required: false,
+		HelpText: fmt.Sprintf(
+			"The log level that Kurtosis itself should output at (%v)",
+			strings.Join(logrus_log_levels.GetAcceptableLogLevelStrs(), "|"),
+		),
+		Default: defaultKurtosisLogLevel,
+		Type:    docker_flag_parser.StringFlagType,
+	},
+}
 
 func main() {
 	// NOTE: we'll want to change the ForceColors to false if we ever want structured logging
@@ -61,10 +87,6 @@ func main() {
 		ForceColors:   true,
 		FullTimestamp: true,
 	})
-
-	// TODO figure out a way to set the loglevel for Kurtosis from here
-
-	// TODO set log level???
 
 	if err := runMain(); err != nil {
 		fmt.Fprintln(logrus.StandardLogger().Out, err)
@@ -74,6 +96,22 @@ func main() {
 }
 
 func runMain() error {
+	flagParser := docker_flag_parser.NewFlagParser(flagConfigs)
+	parsedFlags, err := flagParser.Parse()
+	if err != nil {
+		return stacktrace.Propagate(err, "An error occurred parsing the CLI flags")
+	}
+
+	kurtosisLogLevelStr := parsedFlags.GetString(kurtosisLogLevelArg)
+	kurtosisLogLevel, err := logrus.ParseLevel(kurtosisLogLevelStr)
+	if err != nil {
+		return stacktrace.Propagate(err, "An error occurred parsing Kurtosis loglevel string '%v' to a log level object", kurtosisLogLevelStr)
+	}
+	logrus.SetLevel(kurtosisLogLevel)
+
+	apiContainerImage := parsedFlags.GetString(apiContainerImageArg)
+	jsReplImage := parsedFlags.GetString(javascriptReplImageArg)
+
 	dockerClient, err := client.NewClientWithOpts(client.FromEnv, client.WithAPIVersionNegotiation())
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred creating the Docker client")
@@ -100,6 +138,9 @@ func runMain() error {
 		enclaveId,
 		isPartitioningEnabled,
 	)
+	if err != nil {
+		return stacktrace.Propagate(err, "An error occurred creating an enclave")
+	}
 	defer func() {
 		// Ensure we don't leak enclaves
 		logrus.Info("Removing enclave...")
@@ -112,16 +153,19 @@ func runMain() error {
 		}
 	}()
 
-	logrus.Info("Running REPL...")
-	if err := runReplContainer(dockerManager, enclaveCtx); err != nil {
+	logrus.Debug("Running REPL...")
+	if err := runReplContainer(dockerManager, enclaveCtx, jsReplImage); err != nil {
 		return stacktrace.Propagate(err, "An error occurred running the REPL container")
 	}
-	logrus.Info("REPL exited")
+	logrus.Debug("REPL exited")
 
 	return nil
 }
 
-func runReplContainer(dockerManager *docker_manager.DockerManager, enclaveCtx *enclave_context.EnclaveContext) error {
+func runReplContainer(
+		dockerManager *docker_manager.DockerManager,
+		enclaveCtx *enclave_context.EnclaveContext,
+		javascriptReplImage string) error {
 	enclaveId := enclaveCtx.GetEnclaveID()
 	networkId := enclaveCtx.GetNetworkID()
 	kurtosisApiContainerIpAddr := enclaveCtx.GetAPIContainerIPAddr()